fix: padding-free deny key + daemon-assisted cgroup migration
Two bugs found in integration testing, both fixed:

1. deny_map lookup missed despite matching {dev,ino}: struct ino_key had
   {u32 s_dev; u64 ino} -> 4 padding bytes that differed between the eBPF
   stack key and the daemon's key. Made both fields u64 (no padding). Verified
   via a debug map: cgroup id matched (st_ino == name_to_handle_at == BPF
   cgid), key extraction was correct; padding was the sole cause.

2. owner (the real agent user) could not write cgroup.procs (cgroup v2 needs
   a delegated subtree, and user.slice/user-1000.slice is root-owned here).
   Added daemon-assisted migration: launcher drops a pid-named request in
   /run/agent-sandbox/req; sandboxd (root) inotify-watches it, migrates the
   pid into /sys/fs/cgroup/agent-sandbox, removes the request. Launcher waits
   for the file to disappear, then execs. No setuid; daemon stays the sole
   privilege boundary.

Also: event loop now parses inotify events (req-watch vs denylist-watch);
postinst message path fixed (/usr/bin); test.sh groups check is now
inside-vs-outside (works as root or owner).

make install + scripts/test.sh: ALL PASS (deny ENOENT, cat|pipe + hardlink
blocked, groups/proc/dev transparent).

+++ src/agent_sandbox.h
@@ -7,10 +7,12 @@
  * Keeping this in one place guarantees the deny-map key layout matches exactly.
  */
 
-/* Identifier of a file to deny: {containing-device, inode-number}. */
+/* Identifier of a file to deny: {containing-device, inode-number}.
+ * Both fields are 64-bit so the struct has NO padding — the BPF program and
+ * the daemon must build byte-identical keys for the hash-map lookup. */
 struct ino_key {
-	__u32 s_dev;	/* inode->i_sb->s_dev (== stat st_dev, low 32 bits) */
-	__u64 ino;	/* inode->i_ino (== stat st_ino)                    */
+	__u64 s_dev;	/* inode->i_sb->s_dev (== stat st_dev) */
+	__u64 ino;	/* inode->i_ino (== stat st_ino)       */
 };
 
 #endif /* AGENT_SANDBOX_H */
